% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quint.bootstrapCI.R
\name{quint.bootstrapCI}
\alias{quint.bootstrapCI}
\title{Bootstrap method to compute confidence intervals for Qualitative Interaction Trees (Quint)}
\usage{
quint.bootstrapCI(tree, n_boot, boot_r = 1)
}
\arguments{
\item{tree}{a (pruned) QUINT tree object of class \code{quint}.}

\item{n_boot}{number of bootstrap samples.}

\item{boot_r}{Size of the bootstrap. Default value is 1.}
}
\value{
Returns two list: A first one containing an object of the class \code{quint}, and a list
estimates obtained from the bootstrap procedure containing the following components:
  \item{nleaves}{vector containing the number of leaves in each of the trees obtained by the bootstrap procedure.}
  \item{meanT_1}{a matrix containing the average outcome for treatment 1 in each leaf for each bootstrap.}
  \item{meanT_2}{a matrix containing the average outcome for treatment 2 in each leaf for each bootstrap.}
  \item{meandif}{a matrix containing the difference in means between treatment 1 and treatment 2 in each leaf for each bootstrap sample.}
  \item{bias_est}{vector containing the bias in each leaf of the quint tree.}
  \item{meanboot}{vector containing the bootstrap estimates of the difference of means between treatments in each leaf.}
  \item{CIs}{vector containing the confidence intervals of the estimate of the difference of means between treatments in each leaf.}
  \item{se_est}{vector containing the estimate of the standard error of the difference of means between treatments in each leaf.}
}
\description{
A bootstrap-based algorithm to estimate the confidence intervals of the mean difference
in each leaf obtained by QUINT.
}
\details{
The details of this validaton procedure are described in "Instability of
  QUalitative INteraction Trees: Quantifying uncertainty in decision trees." (
  https://openaccess.leidenuniv.nl/handle/1887/83059)
}
\examples{
data(bcrp)
formula1<- I(cesdt1-cesdt3)~cond | nationality+marital+wcht1+age+
  trext+comorbid+disopt1+uncomt1+negsoct1

set.seed(10)
control1<-quint.control(maxl=5,B=2)
quint1<-quint(formula1, data= subset(bcrp,bcrp$cond<3),control=control1) #Grow a QUINT tree

prquint1<-prune(quint1) #Prune tree to optimal size

bootquint1<-quint.bootstrapCI(prquint1, n_boot = 3) #compute the new confidence intervals
summary(bootquint1$tree) #Summary information of the obtained tree
}
\references{
Dusseldorp E. and Van Mechelen I. (2014). Qualitative interaction trees:
  a tool to identify qualitative treatment-subgroup interactions.
  \emph{Statistics in Medicine, 33}(2), 219-237. DOI: 10.1002/sim.5933.
  Beck C., Dusseldorp E. and Fokkema M. (2019). Instability of
  QUalitative INteraction Trees: Quantifying uncertainty in decision trees.
  (https://openaccess.leidenuniv.nl/handle/1887/83059))
}
\seealso{
\code{\link{quint}}, \code{\link{prune.quint}}, \code{\link{quint.control}}
}
